Strategy to poll a json endpoint for the node list.
